requirements.txt
requests
psycopg2-binary
beautifulsoup4
pandas
scikit-learn
matplotlib
